Extract SwiftEnumDescriptor from SwiftASTContext.cpp
Move the code to a new SwiftEnum.h/cpp in Language/Swift.  SwiftASTContext
did not need to know how SwiftEnum's work in detail, and this prepares us
for using reflection to get the enum details, which really didn't belong
in SwiftASTContext.cpp.
